app/src/main/java/com/thalukanyo/android/coffeorder/models/Customer.java
package com.thalukanyo.android.coffeorder.models;

public class Customer {
    private String customerName;
    private String customerPhoneNumber;
    private String customerEmail;

    public Customer(String customerName, String customerPhoneNumber, String customerEmail) {
        this.customerName = customerName;
        this.customerPhoneNumber = customerPhoneNumber;
        this.customerEmail = customerEmail;
    }

    public boolean isPhoneNumberValid(String number) {
        return (number.length() == 10) ? true : false;
    }

    public String getCustomerName() {
        return customerName;
    }

    public void setCustomerName(String customerName) {
        this.customerName = customerName;
    }

    public String getCustomerPhoneNumber() {
        return customerPhoneNumber;
    }

    public void setCustomerPhoneNumber(String customerPhoneNumber) {
        this.customerPhoneNumber = customerPhoneNumber;
    }

    public String getCustomerEmail() {
        return customerEmail;
    }

    public void setCustomerEmail(String customerEmail) {
        this.customerEmail = customerEmail;
    }

    @Override
    public String toString() {
        return "Customer{" +
                "customerName='" + customerName + '\'' +
                ", customerPhoneNumber='" + customerPhoneNumber + '\'' +
                ", customerEmail='" + customerEmail + '\'' +
                '}';
    }
}
